cc: fix snake_case to camelCase convention
