+++ src/data/brandRepository.js
@@ -1,57 +1,46 @@
 'use strict';
 
 const { defaultsDeep } = require('lodash');
-const BaseRepository = require('./base_repository');
-const CustomerDto = require('./models/Customers');
+const BaseRepository = require('./baseRepository');
+const BrandDto = require('./models/Brands');
 
-const { CollectionModel, CustomerModel } = require('../models');
-const { compareTwoText, hashText } = require('../libs/bcrypt_helper');
+const { CollectionModel, BrandModel } = require('../models');
 const { logger } = require('../libs/logger');
 const { Utils } = require('../libs/utils');
 
 const defaultOpts = {};
 
-class CustomerRepository extends BaseRepository {
-  /**
-   * @param {*} opts
-   * @param {RedisClient} redis
-   */
-  constructor(opts, redis) {
-    super();
-    /** @type {defaultOpts} */
-    this.opts = defaultsDeep(opts, defaultOpts);
-    this.redis = redis;
-  }
+class BrandRepository extends BaseRepository {
   /**
    *
    * @param {*} query
    * @param {Number} limit
    * @param {Number} page
    * @param {Boolean} count with count number of records
-   * @returns {Promise<CollectionModel<CustomerModel>>}
+   * @returns {Promise<CollectionModel<BrandModel>>}
    */
-  async findCustomer(query = {}, limit = 10, page = 1, count = false) {
+  async findBrand(query = {}, limit = 10, page = 1, count = false) {
     const coll = new CollectionModel();
     coll.page = page;
     coll.limit = limit;
     try {
-      const docs = await CustomerDto.find(query)
+      const docs = await BrandDto.find(query)
         .limit(limit)
         .skip((page - 1) * limit)
         .sort({ createdAt: -1 });
       if (docs.length > 0) {
-        coll.data = docs.map((item) => CustomerModel.fromMongo(item));
+        coll.data = docs.map((item) => BrandModel.fromMongo(item));
       }
-      coll.total = count ? await CustomerDto.count(query) : docs.length;
+      coll.total = count ? await BrandDto.count(query) : docs.length;
     } catch (err) {
       logger.error(err, err.message);
     }
     return coll;
   }
   async findAllData(data) {
-    let coll = await CustomerDto.find({ ...data });
+    let coll = await BrandDto.find({ ...data });
     if (coll.length > 0) {
-      coll = coll.map((item) => CustomerModel.fromMongo(item));
+      coll = coll.map((item) => BrandModel.fromMongo(item));
     }
     return coll;
   }
@@ -59,23 +48,23 @@ class CustomerRepository extends BaseRepository {
     if (data == null) {
       return;
     }
-    const doc = await new CustomerDto(data).save();
-    const inserted = CustomerModel.fromMongo(doc);
+    const doc = await new BrandDto(data).save();
+    const inserted = BrandModel.fromMongo(doc);
     return inserted;
   }
   async findOne(key, value) {
-    const coll = await CustomerDto.findOne({ [key]: value });
-    const inserted = CustomerModel.fromMongo(coll);
+    const coll = await BrandDto.findOne({ [key]: value });
+    const inserted = BrandModel.fromMongo(coll);
     return inserted;
   }
   async findData(data) {
-    const docs = await CustomerDto.find(data);
-    const coll = docs.map((item) => CustomerModel.fromMongo(item));
+    const docs = await BrandDto.find(data);
+    const coll = docs.map((item) => BrandModel.fromMongo(item));
     return coll;
   }
   async update(query = {}, update = {}) {
     try {
-      const coll = await CustomerDto.findOneAndUpdate(query, update, {
+      const coll = await BrandDto.findOneAndUpdate(query, update, {
         new: true,
       });
       return coll;
@@ -85,7 +74,7 @@ class CustomerRepository extends BaseRepository {
   }
   async updateMany(query = {}, update = {}) {
     try {
-      const coll = await CustomerDto.updateMany(query, update, {
+      const coll = await BrandDto.updateMany(query, update, {
         new: true,
       });
       return coll;
@@ -93,7 +82,7 @@ class CustomerRepository extends BaseRepository {
       logger.error(err, err.message);
     }
   }
-  async updateCustomerById(msg) {
+  async updateBrandById(msg) {
     const { uid, data } = msg;
     const coll = await this.update(
       { uid: uid },
@@ -101,17 +90,17 @@ class CustomerRepository extends BaseRepository {
         ...data,
       },
     );
-    const inserted = CustomerModel.fromMongo(coll);
+    const inserted = BrandModel.fromMongo(coll);
     return inserted;
   }
   async delete(data) {
     if (data == null) {
       return;
     }
-    const coll = await CustomerDto.delete({ uid: data });
+    const coll = await BrandDto.delete({ uid: data });
     return coll;
   }
-  async deleteCustomerById(value) {
+  async deleteBrandById(value) {
     const deleted = await this.delete(value);
     return deleted;
   }
@@ -120,15 +109,10 @@ class CustomerRepository extends BaseRepository {
     if (value == null) {
       return;
     }
-    const coll = await CustomerDto.delete({ [key]: { $in: value } });
+    const coll = await BrandDto.delete({ [key]: { $in: value } });
     return coll;
   }
-  async generateCode() {
-    const count = await CustomerDto.find();
-    const total = count.length + 1;
-    const number = ('0000' + total).slice(-4);
-    return `KH${number}`;
-  }
+  ////
   async search(data) {
     const paging = {
       total: 0,
@@ -159,15 +143,20 @@ class CustomerRepository extends BaseRepository {
       {
         $project: {
           _id: 0,
+          uid: 1,
+          code: 1,
+          name: 1,
+          status: 1,
+          createdAt: 1,
         },
       },
     ];
-    const coll = await CustomerDto.aggregate(pipe)
+    const coll = await BrandDto.aggregate(pipe)
       .sort({ createdAt: -1 })
       .skip((data.page - 1) * data.limit)
       .limit(data.limit);
 
-    const total = await CustomerDto.aggregate(pipe).count('code');
+    const total = await BrandDto.aggregate(pipe).count('code');
     paging.total = total.length > 0 ? total[0].code : 0;
 
     if (coll.total === 0) {
@@ -175,30 +164,5 @@ class CustomerRepository extends BaseRepository {
     }
     return [coll, paging];
   }
-  async comparePasswordLogin(data) {
-    const coll = await this.findCustomer(
-      {
-        // $or: [
-        //   { clientCode: { $regex: `^${username}$`, $options: 'i' } },
-        //   { email: { $regex: `^${username}$`, $options: 'i' } },
-        // ],
-        // ...role,
-        username: { $regex: `^${data.username}$`, $options: 'i' },
-      },
-      1,
-      1,
-      false,
-    );
-    if (coll.total === 0) {
-      return null;
-    }
-    const customer = coll.data[0];
-    const passwordFind = customer.password;
-    if (!compareTwoText(data.password, passwordFind)) {
-      return null;
-    }
-    customer.password = undefined;
-    return customer;
-  }
 }
-module.exports = CustomerRepository;
+module.exports = BrandRepository;
