+++ src/data/productTypeRepository.js
@@ -1,70 +1,80 @@
 'use strict';
 
 const { defaultsDeep } = require('lodash');
-const BaseRepository = require('./base_repository');
-const BrandDto = require('./models/Brands');
+const BaseRepository = require('./baseRepository');
+const ProductTypeDto = require('./models/ProductTypes');
 
-const { CollectionModel, BrandModel } = require('../models');
+const { CollectionModel, ProductTypeModel } = require('../models');
 const { logger } = require('../libs/logger');
 const { Utils } = require('../libs/utils');
 
 const defaultOpts = {};
 
-class BrandRepository extends BaseRepository {
+class ProductTypeRepository extends BaseRepository {
   /**
    *
    * @param {*} query
    * @param {Number} limit
    * @param {Number} page
    * @param {Boolean} count with count number of records
-   * @returns {Promise<CollectionModel<BrandModel>>}
+   * @returns {Promise<CollectionModel<ProductTypeModel>>}
    */
-  async findBrand(query = {}, limit = 10, page = 1, count = false) {
+  async findProductType(query = {}, limit = 10, page = 1, count = false) {
     const coll = new CollectionModel();
     coll.page = page;
     coll.limit = limit;
     try {
-      const docs = await BrandDto.find(query)
+      const docs = await ProductTypeDto.find(query)
         .limit(limit)
         .skip((page - 1) * limit)
         .sort({ createdAt: -1 });
       if (docs.length > 0) {
-        coll.data = docs.map((item) => BrandModel.fromMongo(item));
+        coll.data = docs.map((item) => ProductTypeModel.fromMongo(item));
       }
-      coll.total = count ? await BrandDto.count(query) : docs.length;
+      coll.total = count ? await ProductTypeDto.count(query) : docs.length;
     } catch (err) {
       logger.error(err, err.message);
     }
     return coll;
   }
   async findAllData(data) {
-    let coll = await BrandDto.find({ ...data });
+    let coll = await ProductTypeDto.find({ ...data });
     if (coll.length > 0) {
-      coll = coll.map((item) => BrandModel.fromMongo(item));
+      coll = coll.map((item) => ProductTypeModel.fromMongo(item));
     }
     return coll;
+  }
+  async create(data) {
+    if (data == null) {
+      return;
+    }
+    const doc = await ProductTypeDto.insertMany(data);
+    if (doc != null) {
+      return true;
+    }
   }
   async createOne(data) {
     if (data == null) {
       return;
     }
-    const doc = await new BrandDto(data).save();
-    const inserted = BrandModel.fromMongo(doc);
+    const doc = await new ProductTypeDto(data).save();
+    const inserted = ProductTypeModel.fromMongo(doc);
     return inserted;
   }
   async findOne(key, value) {
-    const coll = await BrandDto.findOne({ [key]: value });
-    const inserted = BrandModel.fromMongo(coll);
+    const coll = await ProductTypeDto.findOne({ [key]: value });
+    const inserted = ProductTypeModel.fromMongo(coll);
     return inserted;
   }
   async findData(data) {
-    const docs = await BrandDto.find(data);
-    const coll = docs.map((item) => BrandModel.fromMongo(item));
+    const docs = await ProductTypeDto.find(data);
+    const coll = docs.map((item) => ProductTypeModel.fromMongo(item));
     return coll;
   }
+
   async update(query = {}, update = {}) {
     try {
-      const coll = await BrandDto.findOneAndUpdate(query, update, {
+      const coll = await ProductTypeDto.findOneAndUpdate(query, update, {
         new: true,
       });
       return coll;
@@ -74,7 +84,7 @@ class BrandRepository extends BaseRepository {
   }
   async updateMany(query = {}, update = {}) {
     try {
-      const coll = await BrandDto.updateMany(query, update, {
+      const coll = await ProductTypeDto.updateMany(query, update, {
         new: true,
       });
       return coll;
@@ -82,7 +92,7 @@ class BrandRepository extends BaseRepository {
       logger.error(err, err.message);
     }
   }
-  async updateBrandById(msg) {
+  async updateProductTypeById(msg) {
     const { uid, data } = msg;
     const coll = await this.update(
       { uid: uid },
@@ -90,17 +100,17 @@ class BrandRepository extends BaseRepository {
         ...data,
       },
     );
-    const inserted = BrandModel.fromMongo(coll);
+    const inserted = ProductTypeModel.fromMongo(coll);
     return inserted;
   }
   async delete(data) {
     if (data == null) {
       return;
     }
-    const coll = await BrandDto.delete({ uid: data });
+    const coll = await ProductTypeDto.delete({ uid: data });
     return coll;
   }
-  async deleteBrandById(value) {
+  async deleteProductTypeById(value) {
     const deleted = await this.delete(value);
     return deleted;
   }
@@ -109,10 +119,9 @@ class BrandRepository extends BaseRepository {
     if (value == null) {
       return;
     }
-    const coll = await BrandDto.delete({ [key]: { $in: value } });
+    const coll = await ProductTypeDto.delete({ [key]: { $in: value } });
     return coll;
   }
-  ////
   async search(data) {
     const paging = {
       total: 0,
@@ -151,12 +160,12 @@ class BrandRepository extends BaseRepository {
         },
       },
     ];
-    const coll = await BrandDto.aggregate(pipe)
+    const coll = await ProductTypeDto.aggregate(pipe)
       .sort({ createdAt: -1 })
       .skip((data.page - 1) * data.limit)
       .limit(data.limit);
 
-    const total = await BrandDto.aggregate(pipe).count('code');
+    const total = await ProductTypeDto.aggregate(pipe).count('code');
     paging.total = total.length > 0 ? total[0].code : 0;
 
     if (coll.total === 0) {
@@ -164,5 +173,25 @@ class BrandRepository extends BaseRepository {
     }
     return [coll, paging];
   }
+  async listProductType() {
+    const pipe = [
+      {
+        $match: { status: true },
+      },
+      {
+        $project: {
+          _id: 0,
+          uid: 1,
+          code: 1,
+          name: 1,
+          createdAt: 1,
+        },
+      },
+    ];
+    const coll = await ProductTypeDto.aggregate(pipe).sort({
+      createdAt: -1,
+    });
+    return coll;
+  }
 }
-module.exports = BrandRepository;
+module.exports = ProductTypeRepository;
