+++ src/domain/product.js
@@ -1,12 +1,12 @@
 'use strict';
 /**
  * @typedef {import("./policy")} PolicyService
- * @typedef {import("../data/product_repository")} productRepository
- * @typedef {import("../data/customer_repository")} customerRepository
- * @typedef {import("../data/order_repository")} orderRepository
- * @typedef {import("../data/user_repository")} userRepository
- * @typedef {import("../data/productType_repository")} productTypeRepository
- * @typedef {import("../data/brand_repository")} brandRepository
+ * @typedef {import("../data/productRepository")} productRepository
+ * @typedef {import("../data/customerRepository")} customerRepository
+ * @typedef {import("../data/orderRepository")} orderRepository
+ * @typedef {import("../data/userRepository")} userRepository
+ * @typedef {import("../data/productTypeRepository")} productTypeRepository
+ * @typedef {import("../data/brandRepository")} brandRepository
  */
 const { defaultsDeep } = require('lodash');
 const { ulid } = require('ulid');
